+++ src/rendering/raylib/RaylibRenderer.cpp
@@ -40,10 +40,13 @@ void RaylibRender::ClearImpl(const cnf::Color &color) const noexcept {
 }
 void RaylibRender::CircleImpl(cnf::PosType x, cnf::PosType y, float r,
                               const cnf::Color &color) const noexcept {
-  DrawCircle(x, y, r, Color(color.r, color.g, color.b, color.a));
+  DrawCircle(static_cast<int>(x), static_cast<int>(y), static_cast<float>(r),
+             Color(color.r, color.g, color.b, color.a));
 }
 void RaylibRender::RectImpl(cnf::PosType x, cnf::PosType y, cnf::PosType width,
-                           cnf::PosType height,
-                           const cnf::Color &color) const noexcept {
-  DrawRectangle(x, y, width, height, Color(color.r, color.g, color.b, color.a));
+                            cnf::PosType height,
+                            const cnf::Color &color) const noexcept {
+  DrawRectangle(static_cast<int>(x), static_cast<int>(y),
+                static_cast<int>(width), static_cast<int>(height),
+                Color(color.r, color.g, color.b, color.a));
 }
